backend/requirements.txt
crewai[google-genai]==1.7.2
openai==1.83.0
crewai-tools
langchain-google-genai
google-generativeai
pinecone-client
langchain-pinecone
opencv-python
mediapipe
numpy>=2.1.0
python-dotenv
streamlit
langchain-groq
groq
litellm==1.34.0
fastapi
uvicorn
pandas
